Create 数组（Find the Two Largest Numbers）
The process of find the two largest values of the 10 numbers. [Note: You may input each number only once.]
输入格式:
Inputs a series of 10 integer numbers in a line. Seperate each number by a space.
输出格式:
Prints the largest two of the numbers in a line. Seperate each number by a space.
输入样例:
77 88 22 33 11 2 8 7 9 89

输出样例:
89 88